4.21 Blog Deletion with token

+++ part4/bloglist/controllers/blogs.js
@@ -44,6 +44,11 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).json('Deleted the entry!')
 })
